Sort search results
**via invidious API**

Adds radio buttons to sort search results by
- rating
- relevance (like before)
- upload date
- view count

The radio buttons are hidden by default and will be shown when checking the settings button next to the search field.

The "rating" is not correct (in invidious): it sorts from low to high.

![grafik](/uploads/50f0eef58536f544f87cbaa3d31b34a9/grafik.png)

Extends qinvidious to support all parameters of the invidious search API.