Improve video loading appearance
Previously, when loading new videos the experience isn't the greatest.
The thumbnail may take some time to load, and there's no clear
indication if the video is loading or just paused. This gets worse if
you are starting a new video, because the last video's still frame is
visible while the next video loads.

Now, there's a busy indicator during loading and the video player is
hidden when stopped, so it stops displaying the last frame.