Re-organize the current navigation tabs into four categories
Previously there was "Videos" and then "Search". This is pretty
confusing as "Videos" encompassed "Popular", "Trending", "Subscriptions"
and so on.

Now there is a base "FeedPage" which does the same thing the old page
did, but separate each function into their own pages. There is now a
"Popular" page for the popular videos in your instance (previously
called "Top"). Below that is Trending and all of its own subtopics. The
last page is Subscriptions, which is only available if logged in.