Begin moving some video controller logic into C++
Created a new VideoController class, which dictates some of the video
player and synchronization logic in C++. This will be useful in the
future once queueing and playlist support is added, and we can keep that
stuff out of JavaScript or the MpvObject class.

Right now it handles opening and managing the two player states (normal
and picture in picture) and attempts to synchronize the player controls
between them.