Remove popular page
There is numerous issues with the popular page. First, it's the default
page for Invidious instances, but it happens to be disabled by most
administrators. There's no way to tell through the API if the page is
disabled or there is an actual error, so it results in terrible UX.

Also, the concept of a "popular on this instance" page only exists in
Invidious. The trending page is now the new default page, which is much
more reliable.