Reduce the amount of the video title duplication
Previously, the video title could be duplicated on screen up to four
times (!!) In the window title, the page title, in the video player and
then below the video player.

Most of these are extraneous, so the video name in the page title
is removed. If the video title inside the player is removed if it's not
picture-in-picture.