Overhaul login and credentials system
The old system and architecture was tied to Invidious, and specifically
only storing the cookie value. This won't scale at all, so now the API
implementation is in charge of saving and storing any secrets it needs
and not VideoSource.

Standard methods for interacting with the system
keychain are now built into AbstractApi.