Add an error page to the video player, add a way out
Currently Invidious servers are being hammered by Google. This creates
a terrible experience in PlasmaTube if you happen to use one of these
servers, as videos load forever with no clear messaging.

Now PlasmaTube should display the error message, block the rest of the
video page, and show a button to visit the video in your web browser
instead.