Overhaul the server selection UX
Instead of a simple list of servers, the UX of picking a server is
greatly improved. First of all, the servers now exist in a proper
scrollable list view. Secondly, there is now more information available
depending on the video source such as server name, region and icon. A
loading and "no servers found" message is added as well.